src/api_server/requirements.txt
fastapi==0.61.2
uvicorn==0.12.2
uvloop==0.14.0
numpy==1.19.4
pillow==8.0.1
matplotlib==3.3.3
ipython==7.19.0

# do not install pandas by pip, use the apt:
# apt-get install python3-pandas
# virtualenv --system-site-packages -p /usr/bin/python3.8 /opt/venv3.8/
